+++ tests/integration/database_sql_numeric_args_test.go
@@ -11,12 +11,15 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table/query"
 )
 
 func TestDatabaseSqlNumericArgs(t *testing.T) {
 	scope := newScope(t)
 	db := scope.SQLDriverWithFolder(ydb.WithAutoBind(
-		ydb.BindNumeric().WithTablePathPrefix(scope.Folder()),
+		query.TablePathPrefix(scope.Folder()),
+		query.Declare(),
+		query.Numeric(),
 	))
 	dt := time.Date(2023, 3, 1, 16, 34, 18, 0, time.UTC)